switching back to Firebase Authentication
