+++ app/src/main/java/com/loc8r/seattle/mongodb/MongoDBManager.java
@@ -605,8 +605,6 @@ public class MongoDBManager {
         tempPOIList.add(new POI("Sea Monster Lounge",CreateLocation(47.661542, -122.332299),"We feature the very best in Seattle live funk, soul and jazz music, blues/electronic, etc. Tuesday through Sunday.  and we have a covered outdoor patio! it's also a famous spot for musicians getting off gigs elsewhere and wanting to drop in late for a jam session (and some liquor) and you will know why the moment you walk in. Charlie Hunter, Macklemore, Alan Stone, Roy Hargrove, Blake Lewis, Kevin Sawka, etc... 10 NW beers on tap, house liquor infusions, snacks, games, fun!", "https://www.zaske.com/loc8r/seattle_needle.jpg", "street art", "StampIS2","Stamp Text 2"));
         tempPOIList.add(new POI("Library",CreateLocation(47.661173, -122.338994),"The Wilmot Memorial Library opened Sept. 10, 1949, in a house donated by Alice Wilmot Dennis in memory of her sister, Florence Wilmot Metcalf. The branch moved to remodeled quarters in a former police and fire station in 1985 and was renamed the Wallingford-Wilmot Library.\n" +
                 "\n" +
-                " \n" +
-                "\n" +
                 "The branch shared its new space with the 45th Street Clinic, which eventually expanded and asked the library to relocate.", "https://www.zaske.com/loc8r/seattle_library.jpg", "street art", "StampIS3","Stamp Text 3"));
         tempPOIList.add(new POI("Portage Bay Cafe",CreateLocation(47.657846, -122.317634),"A tasty place to go for an all-American, local/organic/sustainable-focused breakfast or brunch, including cage-free eggs from Olympia, hams and sausages made in-house with Carlton Farms pork, all-organic and local-when-possible veggies, and so forth. Reservations are highly recommended on the weekends.", "https://www.zaske.com/loc8r/seattle_skyline1.jpg", "street art", "StampIS4","Stamp Text 4"));
         tempPOIList.add(new POI("University Barbershop",CreateLocation(47.658945, -122.313323),"A great neighborhood barbershop, serving 'The Ave' for over 75 years", "https://www.zaske.com/loc8r/seattle_market.jpg", "street art", "StampIS5","Stamp Text 5"));
