--- app/src/main/res/layout/activity_map.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout
-xmlns:android="http://schemas.android.com/apk/res/android"
-xmlns:mapbox="http://schemas.android.com/apk/res-auto"
-xmlns:tools="http://schemas.android.com/tools"
-android:id="@+id/map_activity_framelayout"
-android:layout_width="match_parent"
-tools:ignore="MissingPrefix"
-android:layout_height="match_parent"
-tools:context="com.loc8r.seattle.activities.MapActivity">
-
-<com.mapbox.mapboxsdk.maps.MapView
-    android:id="@+id/mapView"
-    mapbox_styleUrl="@string/mapbox_style_light"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginBottom="0dp"
-    android:layout_marginLeft="0dp"
-    android:layout_marginRight="0dp"
-    android:layout_marginTop="0dp"
-    mapbox:mapbox_cameraTargetLat="47.620628"
-    mapbox:mapbox_cameraTargetLng="-122.349031"
-    mapbox:mapbox_cameraTilt="0"
-    mapbox:mapbox_cameraZoom="15"
-    mapbox:mapbox_cameraZoomMin="16"
-    mapbox:mapbox_uiAttribution="false"/>
-
-<android.support.v7.widget.RecyclerView
-    android:id="@+id/map_layout_rv"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="bottom"
-    android:layout_marginBottom="30dp"/>
-
-</FrameLayout>
