+++ app/src/main/res/layout/map_drawer.xml
@@ -20,14 +20,12 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent">
 
-        <android.support.constraint.ConstraintLayout
-            android:id="@+id/hide_click_area"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="8dp"
-            android:layout_marginStart="8dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            tools:layout_editor_absoluteX="8dp"
+            tools:layout_editor_absoluteY="8dp">
 
             <TextView
                 android:id="@+id/textView"
@@ -47,52 +45,57 @@
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@+id/textView"
                 app:srcCompat="@drawable/ic_keyboard_arrow_down_black_24dp" />
-        </android.support.constraint.ConstraintLayout>
 
-        <android.support.constraint.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            tools:layout_editor_absoluteX="0dp"
-            tools:layout_editor_absoluteY="100dp"
-            app:layout_constraintTop_toBottomOf="@+id/hide_click_area">
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:orientation="horizontal">
 
-            <ImageView
-                android:id="@+id/icon_imageView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="16dp"
-                android:layout_marginTop="8dp"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/arrow_down_imageView"
-                app:srcCompat="@drawable/ice_cream_icon" />
+                <ImageView
+                    android:id="@+id/icon_imageView"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_margin="8dp"
+                    app:srcCompat="@drawable/ice_cream_icon" />
 
-            <TextView
-                android:id="@+id/title_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="8dp"
-                android:text="POI Title can be long"
-                app:layout_constraintStart_toEndOf="@+id/icon_imageView"
-                app:layout_constraintTop_toTopOf="@+id/icon_imageView" />
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:orientation="vertical">
 
-            <TextView
-                android:id="@+id/desc_textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="8dp"
-                android:text="A generic description of the POI"
-                app:layout_constraintStart_toStartOf="@+id/title_textView"
-                app:layout_constraintTop_toBottomOf="@+id/title_textView" />
+                    <TextView
+                        android:id="@+id/title_textView"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:text="POI Title can be long"
+                        app:layout_constraintStart_toEndOf="@+id/icon_imageView"
+                        app:layout_constraintTop_toTopOf="@+id/icon_imageView" />
 
-            <Button
-                android:id="@+id/poi_details_button"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="16dp"
-                android:text="DETAILS"
-                app:layout_constraintStart_toStartOf="@+id/title_textView"
-                app:layout_constraintTop_toBottomOf="@+id/desc_textView" />
-        </android.support.constraint.ConstraintLayout>
+                    <TextView
+                        android:id="@+id/desc_textView"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="8dp"
+                        android:layout_weight="1"
+                        android:ellipsize="end"
+                        android:maxLines="3"
+                        android:text="A generic description of the POI"
+                        app:layout_constraintStart_toStartOf="@+id/title_textView"
+                        app:layout_constraintTop_toBottomOf="@+id/title_textView"
+                        app:layout_constraintWidth_default="wrap" />
+
+                    <Button
+                        android:id="@+id/poi_details_button"
+                        android:layout_width="match_parent"
+                        android:layout_height="wrap_content"
+                        android:layout_margin="8dp"
+                        android:text="DETAILS"
+                        app:layout_constraintStart_toStartOf="@+id/title_textView"
+                        app:layout_constraintTop_toBottomOf="@+id/desc_textView" />
+                </LinearLayout>
+
+            </LinearLayout>
+        </LinearLayout>
 
     </android.support.constraint.ConstraintLayout>
